21-12-31  15:00 입양공고 상세보기 수정
1. Controller : 서비스에 getPet 메소드를 추가했습니다.
- 추가) [PetMapper.java](http://PetMapper.java) (getPet 메소드를 추가)
- 추가) PetMapper.xml  (사용하는 db를 to_pet으로 적용)
- 추가) [AddPerService.java](http://AddPerService.java) (getPet 메소드를 추가)
- 삭제) AddPetController 일부 불필요한 부분 삭제(homepage)
1. View
- 수정) index.jsp 본인파트 수정(링크 변경)
- 수정) WaitingPet.jsp 본인파트 수정(Pet 클래스 불러오기)

+++ Together/src/main/java/com/pet/Together/AddPet/Pet.java
@@ -6,7 +6,7 @@ public class Pet {
 	private String breed;
 	private int sex;
 	private int age;
-	private long weight;
+	private double weight;
 	private String info;
 	private String location;
 	private int state;
@@ -16,7 +16,7 @@ public class Pet {
 		super();
 	}
 
-	public Pet(int id, String name, String breed, int sex, int age, long weight, String info, String location,
+	public Pet(int id, String name, String breed, int sex, int age, double weight, String info, String location,
 			int state, Img imgs) {
 		super();
 		this.id = id;
@@ -71,11 +71,11 @@ public class Pet {
 		this.age = age;
 	}
 
-	public long getWeight() {
+	public double getWeight() {
 		return weight;
 	}
 
-	public void setWeight(long weight) {
+	public void setWeight(double weight) {
 		this.weight = weight;
 	}
